requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0

pydantic==2.9.2
pydantic-settings==2.5.2

langchain==0.3.5
langchain-core==0.3.12
langchain-community==0.3.5

langgraph==0.2.32
langgraph-checkpoint==1.0.12

langchain-groq==0.1.4
groq==0.9.0

openai==1.51.0

python-dotenv==1.0.1
duckduckgo-search==6.3.5
